Namespace/project refactor
This aims to reduce the confusion over the namespace by de-emphasizing
the root namespace of RavenUWP. Now the root namespace will be "Sentry"
so the developer knows which service they're using.

This also removes the unnecessary test application, and reorganizes
project components to shared projects. This, along with some new
interfaces will make it easier to share code with other C# client
applications like Xamarin.